requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
psycopg2-binary
sqlalchemy
passlib[bcrypt]
firebase-admin
python-jose[cryptography]

